refactor(Game): add domain package
