+++ src/main/java/bowling/domain/Game.java
@@ -1,4 +1,4 @@
-package bowling;
+package bowling.domain;
 
 public class Game {
     private int[] rolls = new int[21];
